rename ordnance_survey to place
