+++ .secrets.baseline
@@ -142,13 +142,14 @@
         "is_secret": false
       }
     ],
-    "app/internal/ordnance_survey/client.go": [
+    "app/internal/place/client.go": [
       {
         "type": "Secret Keyword",
-        "filename": "app/internal/ordnance_survey/client.go",
+        "filename": "app/internal/place/client.go",
         "hashed_secret": "fca71afec681b7c2932610046e8e524820317e47",
         "is_verified": false,
-        "line_number": 76
+        "line_number": 76,
+        "is_secret": false
       }
     ],
     "app/internal/random/random.go": [
@@ -242,5 +243,5 @@
       }
     ]
   },
-  "generated_at": "2022-10-12T13:58:34Z"
+  "generated_at": "2022-10-13T15:07:03Z"
 }
