+++ app/internal/place/client_test.go
@@ -1,4 +1,4 @@
-package ordnance_survey
+package place
 
 import (
 	"context"
